requirements.txt
PyYAML >= 6.0.2
blue-yonder >= 0.0.24
wild-blue-yonder >= 0.0.5
pillow >= 11.0.0